VP-421: Moved property update logic to BulkUpdatePropertyManager;
