+++ VirtoCommerce.CatalogModule.Data/BulkUpdate/Model/Actions/UpdateProperties/UpdatePropertiesResult.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace VirtoCommerce.CatalogModule.Data.BulkUpdate.Model.Actions.UpdateProperties
+{
+    public class UpdatePropertiesResult : BulkUpdateActionResult
+    {
+    }
+}
