+++ VirtoCommerce.CatalogModule.Data/VirtoCommerce.CatalogModule.Data.csproj
@@ -106,8 +106,9 @@
     </Compile>
     <Compile Include="BulkUpdate\Extensions\BulkUpdateActionDefinitionBuilderExtensions.cs" />
     <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\UpdatePropertiesActionData.cs" />
-    <Compile Include="BulkUpdate\Services\BulkUpdatePropertyManager.cs" />
-    <Compile Include="BulkUpdate\Services\IBulkUpdatePropertyManager.cs" />
+    <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\UpdatePropertiesResult.cs" />
+    <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\BulkUpdatePropertyManager.cs" />
+    <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\IBulkUpdatePropertyManager.cs" />
     <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\UpdatePropertiesActionContext.cs" />
     <Compile Include="BulkUpdate\Model\Actions\UpdateProperties\UpdatePropertiesBulkUpdateAction.cs" />
     <Compile Include="BulkUpdate\Model\ListEntryDataQuery.cs" />
