+++ VirtoCommerce.CatalogModule.Web/Localizations/en.VirtoCommerce.Catalog.json
@@ -464,6 +464,9 @@
           "placeholder": "Add number",
           "description": "Bounds will be sorted automatically before saving. Price ranges are constructed as follows: Under [bound 1], [bound 1] to [bound 2], ... , Over [bound N]"
         }
+      },
+      "action-progress": {
+        "title":  "Excecution progress"
       }
     },
     "widgets": {
